Convert main.c, mrinfo.c and mapper.c to use getopt() and a usage() fn.
This commit also imports the OpenBSD strtonum() function, obsoleting the
local get_number() function.

After a lot of deciphering I've also concluded that the indentation style
used is very similar to the ellemtel C indentation style in Emacs, but with
an offset of 4 instead of 3.  So I have added a Local Variables section at
the end of each C-file, so anyone editing the code with Emacs should now get
the correct behavior.