Check malloc() error, prevent stack overflow and out of bounds array access
This changeset started out as a simple reflow of log & debug messages.
However, an unchecked malloc() return value was found and fixed.  Also, a
potentially huge stack overflow (on archs with fixed stacks) was fixed and in
the same function a potential out of bounds array access was fixed.

    accept_probe: Add missing check of malloc() return value.

    accept_report: Move rt[4096] to heap instead of stack.  Use MAX_NUM_RT
                   macro instead of 4096 and make sure the for() loop doesn't
                   overstep the boundaries on the rt[] array when traversing
                   the incoming route report packet.

Fix missing check for malloc() return value