Bugfix: k_del_vif() does not work properly in Linux
When some interface (known by mrouted) goes down, mrouted tries to remove
related VIF by calling stop_vif(), which in turn calls k_del_vif().  After
k_del_vif() is called, mrouted exits with the following error:

    setsockopt MRT_DEL_VIF on vif 3: Invalid argument

The reason for this is due to differences in the Linux and *BSD MRT_DEL_VIF
API.  The Linux kernel expects to receive a struct vifctl associated with the
VIF to be deleted, *BSD systems on the other hand expect to receive the index
of that VIF.

Signed-off-by: Dan Kruchinin <dkruchinin@acm.org>

 defs.h |    2 +-
 kern.c |   43 +++++++++++++++++++++++++++++++++----------
 vif.c  |    2 +-
 3 files changed, 35 insertions(+), 12 deletions(-)

Also, update ChangeLog and prepare for new release.