Port FreeBSD fix for f72ae3a: mtrace abs() fix with Clang 3.5
The following is the commit log from the FreeBSD repository as noted in
[bug #196166](https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=196166):

Fix net/mrouted build (mtrace) with clang 3.5.0

The net/mrouted port fails to compile with clang 3.5.0, due to the
following -Werror warning:

    mtrace.c:923:12: error: taking the absolute value of unsigned type 'unsigned int' has no effect [-Werror,-Wabsolute-value]
            if (*s || abs(ntohl(n->tr_vifout) - ntohl(p->tr_vifout)) > 100000) {
                      ^
    mtrace.c:923:12: note: remove the call to 'abs' since unsigned values cannot be negative
            if (*s || abs(ntohl(n->tr_vifout) - ntohl(p->tr_vifout)) > 100000) {
                      ^~~
    1 error generated.

This is a bug in the code, because the return value of `ntohl()` is
unsigned, and the difference of two unsigned quantities is still
unsigned.  However, the intent of the expression is to look at the
absolute difference between the two quantities.

Therefore, similar to head
[r274898](https://svnweb.freebsd.org/base?view=revision&revision=274898)
for rtadvd, introduce a small static function that clarifies the intent,
and call it instead.

Signed-off-by: Dimitry Andric <dimitry@andric.com>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>