DVMRP report missing subnet (off by one error)
A subnet is missing in the DVMRP report sent to it's neighbor after the
initial probe occurs.  The issue was with the traversal of a double
linked list from end to start.  Once the traversing pointer reaches the
start of routing table for loop (condition) it would break and always
drop the first subnet.

Fixes issue #14.

Signed-off-by: Matthew Weber <matthew.weber@rockwellcollins.com>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>