DVMRP report missing subnet
A subnet is missing in the DVMRP report sent to it's neighbor after
the initial probe occurs.  The issue was with the traversal of a double
linked list from end to start. Once the traversing pointer reaches
the start of routing table for loop (condition) it would break and
always drop the first subnet.

Signed-off-by: Matthew Weber <matthew.weber@rockwellcollins.com>