Initial support for mroutectl
This patch adds initial support for mroutectl, a control tool similar to
pimctl, smcroutectl, and the OpenBSD dvmrpctl tools.  It communicates
with mrouted using a UNIX domain socket and, when finished, replaces the
current `mrouted -r` function as well as SIGUSR1 and SIGUSR2, which
dumps mrouted status in two different files.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>