mrouted: Reduce alloc()/free() calls in timer queue
This patch significantly reduces the amount of calls to calloc()/free()
being done at runtime.  When a timer is needed the new code now looks
for a pre-allocated timer in a free-list, only if no free timers exist
does mrouted allocate a new one.  At startup no free timers exist, so
that's when mosts of the allocation takes place.

A SIGHUP does not free the timers in the free-list, only SIGTERM.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>