Fix #40: automatically add IP alias as altnet
When an interface has multiple addresses one previously had to add them
manually to mrouted.conf as `altnet net/len` for each phyint.  Adding
them as RPF for the DVMRP routing table.

This patch automates that step, so if you have multiple addresses, and
the base interface is enabled, each secondary IP address is added as an
altnet to mrouted.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>