mroutectl: get width in terminal, under watch(1), and serial console
The TIOCWINSZ ioctl() works on real terminals.  When that fails we may
be running under watch(1), which updates COLUMNS in the environment for
us.  For serial consoles, however, we likely get nothing back, so we use
the VT100 trick to move the cursor waaaay off and the query its actual
position to get the size.  If all fails we fall back to 79 chars, which
is out of courtesy to the old Windows HyperTerminal which messes up any
characters printed in the 80th column and just wraps to the next line.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>