timer: protect against stopping other timers from timer callbacks
This patch is a safety net for timer_age_queue().  It protects against
timer callbacks deleting/clearing other timers.  In particular if two
timers have the same timeout and execute immediately after each other,
the preceding timer may clear the next, in which case we do not want
to call its callback.

The next patches in this series add timer "cleanup" to a callback, which
mandates this change.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>