Properly track lower-version group membership reports, RFC3376
According to RFC3376 we must track lower-version group memberships to
be backwards compatible with lower-version hosts.

When a lower-version host subscribes to a group, this disables any and
all higher-version protocol functions for that group:

- When a group is in IGMPv2 compat mode:
  - Ignore IGMPv3 BLOCK()
  - Ignore IGMPv3 TO_EX({x})

- Additionally, when a group is in IGMPv1 compat mode:
  - IGMPv2 LEAVE
  - IGMPv3 TO_IN()

This patch replaces the age_old_hosts() and alug_old member, which
only handled IGMPv1.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>