Fix #31: Support for `phyint static-group GROUP`
This patch adds static multicast routing support to mrouted.  This is
achieved using a method similar to Cisco and Juniper.  A static group
join is configured per interface (ASM only, like Ciso).  This is kept
forever and no external join/leave (IGMPv2 or IGMPv3) is allowed for
such groups, any IGMP reports with these groups are simply parsed and
the groups ignored.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>