Fix regression introduced in ef6ce3f, Feb 21 2001
So, twenty+ years ago I did a refactor of the mrouted linked list
implementation for the built-in RIP protocol (yes unicast routes).
Turns out this introduced an off-by-one error in route reporting
to neighbors that in particular affect IPIP tunnels.

Consider the following, very likely, setup:

          |   .----.             .----.   |
    LAN A |---| R1 |---- ... ----| R2 |---| LAN B
          |   '----'             '----'   |

With only an IPIP tunnel between R1 and R2, the only routes that
can be sent are LAN A from R1 to R2, and LAN B from R2 to R1, but
the off-by-one regression introduced in ef6ce3f casued no routes
to be sent over the tunnel.  So even though, e.g., UDP multicast
from LAN A could be flooded towards R2, R2 would drop it because
it did not know of the LAN from which the sender's IP originated.

This fix restores the ability to send *all* routes, fixing the problem
introduced in ef6ce3f.  A later patch set will do a proper refactor,
this time with tets, probably using the BSD queue.h tailq API.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>