Add Docker container build, you probably don't want to use this
Running a Docker container on Linux means using Linux network namespaces
to "share" interfaces with the container.  Although they are not really
shared, but rather moved to the container's namespace.  This is probably
not what you want, since you wont be able to access those interfaces
from your host.

If someone can come up with something clever here to work around this,
like macvlan interfaces or similar, we're all ears! :)

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>